requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.45.5
blake3==1.0.4
blinker==1.4
blobfile==3.0.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
compressed-tensors==0.9.3
contourpy==1.3.2
cryptography==3.4.8
cupy-cuda12x==13.4.1
cycler==0.12.1
dbus-python==1.2.18
Deprecated==1.2.18
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fairscale==0.4.13
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
fbgemm_gpu==1.2.0
filelock==3.13.1
fire==0.7.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
gguf==0.16.2
googleapis-common-protos==1.70.0
greenlet==1.1.2
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.4
httpcore==1.0.8
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
ImageHash==4.3.2
importlib_metadata==8.0.0
interegular==0.3.3
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kiwisolver==1.4.8
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llguidance==0.7.16
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mistral_common==1.5.4
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
msgspec==0.19.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.75.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==11.0.0
progressbar2==4.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pycountry==24.6.1
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-utils==3.9.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==5.4.1
pyzmq==26.4.0
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.2
SecretStorage==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
ssh-import-id==5.11
starlette==0.46.2
sympy==1.13.1
systemd-python==234
termcolor==3.0.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.4
trash-cli==0.17.1.14
triton==3.1.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.141
ultralytics-thop==2.0.14
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.34.2
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
xgrammar==0.1.18
yarl==1.20.0
zipp==1.0.0
